requirements.txt
Pillow==5.4.1
matplotlib==3.0.2
paho-mqtt==1.4.0
pytest==4.1.1
textX==1.8.0
